requirements.txt
discord.py==2.3.2
pytest==7.4.4
pytest-asyncio==0.21.0
python-dotenv==1.0.0
motor==3.3.1  # Si usas MongoDB
